fix(revert): process/read lines in history file in reverse order
The processing of lines of the fmtna command ensures that files are processed
before their parent directories. Feedback lines are written in the same
order in the history file. For the same reason we ensure that order when
processing lines, we need to process history lines in reverse order.

Same logic when the processing is from the revert subcommand. A history
file is also created in this case. And to apply it properly, we also
need to process lines in reverse.